feat: build and run from Docker Support
Build and run the generator from docker image support added.

The main aim and motivation behind doing this is, because
of a no. of different yeoman generators are there,
which are compatible and works with the different version of yeoman generators
and also might be incompatible with other yeoman generators or versions of yeoman.

Signed-off-by: Pratik Raj <rajpratik71@gmail.com>